% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_apply_excel_formulas.R
\name{apply_excel_formulas}
\alias{apply_excel_formulas}
\title{Apply Excel formula}
\usage{
apply_excel_formulas(agency_id, list)
}
\arguments{
\item{list}{}
}
\value{
A df
}
\description{
Apply Excel formula
}
\author{
Lillian Nguyen
}
