% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_apply_excel_formulas.R
\name{apply_formula_class}
\alias{apply_formula_class}
\title{Apply formula class}
\usage{
apply_formula_class(df, cols)
}
\arguments{
\item{cols}{}
}
\value{
A df
}
\description{
Set a column's class to formula so that Excel interprets the formulas originally written as strings in R.
}
\author{
Lillian Nguyen
}
