% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_set_calcs.R
\name{apply_standard_calcs}
\alias{apply_standard_calcs}
\title{Apply standard calcs}
\usage{
apply_standard_calcs(df)
}
\arguments{
\item{df}{}
}
\value{
A df
}
\description{
Apply standard calcs
}
\author{
Lillian Nguyen
}
