% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_make_chiefs_report.R
\name{calc_chiefs_report}
\alias{calc_chiefs_report}
\title{Calc Chief's report}
\usage{
calc_chiefs_report(df)
}
\arguments{
\item{df}{}
}
\value{
A df
}
\description{
Reorganize data to show agency surpluses/deficits by Object
}
\author{
Lillian Nguyen
}
