% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_make_chiefs_report.R
\name{calc_chiefs_report_totals}
\alias{calc_chiefs_report_totals}
\title{Calc Chief's report totals}
\usage{
calc_chiefs_report_totals(df)
}
\arguments{
\item{df}{}
}
\value{
A df
}
\description{
Add total row to Chief's report
}
\author{
Lillian Nguyen
}
