% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_import_export.R
\name{export_analyst_calcs}
\alias{export_analyst_calcs}
\title{Export analyst calculations}
\usage{
export_analyst_calcs(df)
}
\arguments{
\item{df}{}
}
\value{
An Excel file with
}
\description{
Create a historical file that will be referenced next quarter to create templates for analysts.
}
\author{
Lillian Nguyen
}
