% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_export.R
\name{export_pivot_tabs}
\alias{export_pivot_tabs}
\title{Export pivot tabs}
\usage{
export_pivot_tabs(agency_id, list)
}
\arguments{
\item{list}{the list provided by subset_agency_data}
}
\value{
Pivot-Object, Pivot-Subobject, and Pivot-Service SurDef tabs
}
\description{
Export pivot tabs
}
\author{
Lillian Nguyen
}
