% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_export.R
\name{export_projections_tab}
\alias{export_projections_tab}
\title{Export projections tab}
\usage{
export_projections_tab(agency_id, list)
}
\arguments{
\item{list}{the list provided by subset_agency_data}
}
\value{
A new Excel file with a Projections tab
}
\description{
Export projections tab
}
\author{
Lillian Nguyen
}
