% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_set_calcs.R
\name{import_analyst_calcs}
\alias{import_analyst_calcs}
\title{Import analyst calculations}
\usage{
import_analyst_calcs()
}
\value{
A df
}
\description{
Import the csv file containing the calculations chosen by analysts for the
previous quarter
}
\author{
Lillian Nguyen
}
