% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_import_export.R
\name{import_analyst_files}
\alias{import_analyst_files}
\title{Import analyst files}
\usage{
import_analyst_files(files)
}
\arguments{
\item{files}{A list of file paths}
}
\value{
An Excel file with
}
\description{
Create a historical file that will be referenced next quarter to create templates for analysts.
}
\author{
Lillian Nguyen
}
