% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_write_excel_formulas.R
\name{make_pivots}
\alias{make_pivots}
\title{Make pivots}
\usage{
make_pivots(df, type, proj = "quarterly")
}
\arguments{
\item{df}{}
}
\value{
A df
}
\description{
Write Excel formulas to mimic Pivot tables
}
\author{
Lillian Nguyen
}
