% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_write_excel_formulas.R
\name{make_proj_formulas}
\alias{make_proj_formulas}
\title{Make projection formulas}
\usage{
make_proj_formulas(df, manual = "zero")
}
\arguments{
\item{df}{}
}
\value{
A df
}
\description{
Write Excel formulas for Projection and Surplus/Deficit as strings
}
\author{
Lillian Nguyen
}
