% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_rename_factor_object.R
\name{rename_factor_object}
\alias{rename_factor_object}
\title{Rename factor object}
\usage{
rename_factor_object(df)
}
\arguments{
\item{df}{}
}
\value{
A df
}
\description{
Shorten object names and convert to ordered factor in preparation for Chief's Report PDF
}
\author{
Lillian Nguyen
}
