% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_make_chiefs_report.R
\name{run_chiefs_report}
\alias{run_chiefs_report}
\title{Run Chief's report}
\usage{
run_chiefs_report(df)
}
\arguments{
\item{df}{}
}
\value{
A PDF
}
\description{
Generates a formatted PDF with agency surpluses/deficits by Object,
for review by the Finance Director and Deputy Finance Director.
}
\author{
Lillian Nguyen
}
