% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_import_export.R
\name{run_summary_reports}
\alias{run_summary_reports}
\title{Run summary reports}
\usage{
run_summary_reports(df)
}
\arguments{
\item{df}{}
}
\value{
An Excel file with citywide projections, and summary "Pivot" tabs by Object, Subobject, and Agency
}
\description{
Run summary reports
}
\author{
Lillian Nguyen
}
