% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_cols}
\alias{setup_cols}
\title{Setup column names}
\usage{
setup_cols(proj = "quarterly")
}
\arguments{
\item{proj}{A string for projection type, either "quarterly" or "monthly"}
}
\value{
A list
}
\description{
Use params to establish column names, so they can easily be referred to dynamically
}
\author{
Lillian Nguyen
}
