% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_internal}
\alias{setup_internal}
\title{Setup internal values}
\usage{
setup_internal(proj = "quarterly")
}
\arguments{
\item{proj}{A string for projection type, either "quarterly" or "monthly"}
}
\value{
A list
}
\description{
Use params to establish values that help with later calculations
}
\author{
Lillian Nguyen
}
