% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_subset.R
\name{subset_agency_data}
\alias{subset_agency_data}
\title{Subset agency data}
\usage{
subset_agency_data(agency_id, proj = "quarterly")
}
\arguments{
\item{agency_id}{A string}

\item{proj}{"quarterly" or "monthly"}
}
\value{
A list
}
\description{
Subset agency data
}
\author{
Lillian Nguyen
}
